userservices/requirements.txt
Django==2.2.3
djangorestframework==3.10.2
djongo==1.2.33
sqlparse==0.2.4
